+++ apollo-portal/src/main/resources/static/scripts/directive/item-modal-directive.js
@@ -41,6 +41,7 @@ function itemModalDirective($translate, toastr, $sce, AppUtil, EventManager, Con
             scope.showHiddenChars = showHiddenChars;
             scope.changeType = changeType;
             scope.validateItemValue = validateItemValue;
+            scope.formatContent = formatContent;
 
             $('#itemModal').on('show.bs.modal', function (e) {
                 scope.showHiddenCharsContext = false;
@@ -242,6 +243,17 @@ function itemModalDirective($translate, toastr, $sce, AppUtil, EventManager, Con
                     return '<mark>#' + $translate.instant('ItemModal.ChineseComma') + '#</mark>';
                 }
 
+            }
+
+            // 格式化
+            function formatContent() {
+                if (scope.showJsonError) {
+                    return;
+                }
+                var raw = scope.item.value;
+                if (scope.item.type === '3') {
+                    scope.item.value = JSON.stringify(JSON.parse(raw), null, 4);
+                }
             }
         }
     }